Refactoring for move method
