+++ velocity/src/main/java/com/alessiodp/parties/velocity/messaging/VelocityPartiesMessageDispatcher.java
@@ -228,8 +228,8 @@ public class VelocityPartiesMessageDispatcher extends VelocityMessageDispatcher 
 	
 	public void sendConfigs() {
 		if (ConfigMain.PARTIES_BUNGEECORD_PACKETS_CONFIG_SYNC) {
-			PartiesPacket packet = makePacket(PartiesPacket.PacketType.CONFIGS)
-					.setConfigData(((PartiesConfigurationManager) plugin.getConfigurationManager()).makePacketConfigData());
+			PartiesPacket packet = makePacket(PartiesPacket.PacketType.CONFIGS);
+			packet.setConfigData(packet.makePacketConfigData());
 			sendPacketToBungeecord(packet);
 		}
 	}
